Removed untyped simona service.
